Check if Conic projection will have cone at polars

Check if Conic projection will have cone at polars

Usage

checkConicOK(lat0, lon0, projectionString, latmax, latmin, lonmax, lonmin)

Arguments

lat0

A numerical value, central latitude of the map extent

lon0

A numerical value, central longitude of the map extent

projectionString

A string, the full name of the projection, should be one of "Lambert conformal conic", "Albers equal area conic" and "Lambert azimuthal equal area"

latmax

A numerical value, max latitude of the map extent

latmin

A numerical value, min latitude of the map extent

lonmax

A numerical value, max longitude of the map extent

lonmin

A numerical value, min longitude of the map extent

Value

A integer value from -1/0/1, 1 means passing the test, -1 means failing to pass the test, 0 means the case of Albers when the fan of the selected extent spans less than 180° around a pole