Funcion for regional maps with an square-shaped extent

Funcion for regional maps with an square-shaped extent

Usage

printSquareFormat(property, center, latmax, latmin, datum, unit)

Arguments

property

Projection property, should be one of "Equalarea", "Conformal" and "Equidistant".

center

A list with with numerical values named with lng & lat element, which are the median longitude and latitude values

latmax

A numerical value, max latitude of the map extent

latmin

A numerical value, min latitude of the map extent

datum

A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)

unit

A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)

Value

A list of strings named with PROJ & WKT