Projection for the area of interest with equidistant projection

Auto selecting equidistant projections based on the geological shape and projection characteristics. Function will show messages of the basis how the projection is selected. Please note that there's a longitudinal (less than 160) and latitudinal (less than 80) range limit. If area of interest larger than this please use proj_hemisphere() or use proj_specify().

Usage

proj_equidistant(obj, output_type = "proj4", datum = "WGS84", unit = "m")

Arguments

obj

Input geo data, should be one of:

An object can be accepted by sf::st_bbox() to compute the bounding box

A named list with longitude and latitude extents with names of "xmin", "xmax" "ymin" and "ymax"

output_type

A string for expected output, either "proj4" or "WKT"

datum

A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)

unit

A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)

Value

A proj4 or WKT string

See also

Examples

proj_equidistant(c(xmax=112,xmin=156,ymin=6,ymax=23))
#> ## The map extent is relatively large, choose projection considering map shape
#> ## Close to equator
#> ## Extent is not crossing equator
#> [1] "+proj=eqc +lon_0=-46 +lat_ts=14.5 +datum=WGS84 +units=m +no_defs"