Generating projection by the geo-boundary and the shorthand name

Generating projection by the geo-boundary and the shorthand name

Usage

proj_specify(
  obj,
  prj,
  output_type = "proj4",
  datum = "WGS84",
  unit = "m",
  x0 = NA_real_,
  k0 = NA_real_
)

Arguments

obj

An object to compute the bounding box from, which can be accepted by sf::st_bbox()

prj

A string representing the shorthand name of the desired projection

output_type

A string for expected output, either "proj4" or "WKT"

datum

A string for the datum used with the coordinates (currently only 'WGS84', 'ETRS89' and 'NAD83' supported)

unit

A string for horizontal coordinate system units (currently only 'm' and 'ft' supported)

x0

A numeric value specifying the longitude offset

k0

A numeric value specifying the scale parameter

Value

A proj4 or WKT string as output_type specified

Examples

proj_specify(sf::st_bbox(c(xmin = -30,ymin = -25,xmax = 20,ymax = 5),crs = 4326), "lcc")
#> [1] "+proj=lcc +lon_0=-5 +lat_1=-20 +lat_2=0 +lat_0=-10 +datum=WGS84 +units=m +no_defs"