Specify a world map projection

Specify a world map projection

Usage

proj_world(world_proj, lng = 0, output_type = "proj4")

Arguments

world_proj

A string of proj shorthand, could be extracted from world_proj_list

lng

A numerical value of the expected central longitude

output_type

A string for expected output, either "proj4" or "WKT"

Value

A proj4 or WKT string

Examples

proj_world(world_proj_list$compromise$barrel_shape$Natural_Earth, 150)
#> Please note that using this projection without 0 as the central longitude may result in unsatisfactory outcomes for raster data reprojections.
#> [1] "+proj=natearth +lon_0=150 +datum=WGS84 +units=m +no_defs"